move architects to external
